+++ packages/components/data-table-manager/src/column-settings-manager/column-settings-manager.tsx
@@ -1,3 +1,4 @@
+// TODO: @redesign cleanup
 import {
   useMemo,
   useCallback,
@@ -13,7 +14,7 @@ import { DragDropContext, type DropResult } from 'react-beautiful-dnd';
 import debounce from 'debounce-promise';
 import differenceWith from 'lodash/differenceWith';
 import styled from '@emotion/styled';
-import { designTokens } from '@commercetools-uikit/design-system';
+import { designTokens, useTheme } from '@commercetools-uikit/design-system';
 import AsyncSelectInput from '@commercetools-uikit/async-select-input';
 import FieldLabel from '@commercetools-uikit/field-label';
 import Spacings from '@commercetools-uikit/spacings';
@@ -138,6 +139,7 @@ export const ColumnSettingsManager = (props: TColumnSettingsManagerProps) => {
   }
 
   const intl = useIntl();
+  const { themedValue } = useTheme();
   const [isDragging, setIsDragging] = useState(false);
   const { searchHiddenColumns } = props;
 
@@ -186,14 +188,17 @@ export const ColumnSettingsManager = (props: TColumnSettingsManagerProps) => {
       containerTheme={props.managerTheme}
     >
       <DragDropContext onDragEnd={handleDragEnd} onDragStart={handleDragStart}>
-        <Spacings.Inline scale="m">
+        <Spacings.Inline scale={themedValue('m', 'l')}>
           <DroppableContainer
             isDragging={isDragging}
             aria-labelledby="hidden-columns"
           >
-            <Spacings.Stack>
-              <Spacings.Inline scale="xs" alignItems="center">
-                <EyeCrossedIcon size="medium" />
+            <Spacings.Stack scale={themedValue('s', 'm')}>
+              <Spacings.Inline
+                scale={themedValue('xs', 's')}
+                alignItems="center"
+              >
+                <EyeCrossedIcon size={themedValue('medium', 'big')} />
                 <FieldLabel
                   id="hidden-columns"
                   title={intl.formatMessage(messages.hiddenColumns)}
@@ -227,7 +232,7 @@ export const ColumnSettingsManager = (props: TColumnSettingsManagerProps) => {
             isDragging={isDragging}
             aria-labelledby="visible-columns"
           >
-            <Spacings.Stack>
+            <Spacings.Stack scale={themedValue('s', 'm')}>
               <Spacings.Inline scale="xs" alignItems="center">
                 <EyeIcon size="medium" />
                 <FieldLabel