+++ packages/components/data-table-manager/src/data-table-manager.tsx
@@ -1,3 +1,4 @@
+// TODO: @redesign cleanup
 import {
   useMemo,
   cloneElement,
@@ -5,6 +6,7 @@ import {
   type ReactNode,
   type MouseEventHandler,
 } from 'react';
+import { useTheme } from '@commercetools-uikit/design-system';
 import Spacings from '@commercetools-uikit/spacings';
 import DataTableSettings, {
   type TDataTableSettingsProps,
@@ -147,6 +149,7 @@ type TDataTableManagerProps = {
 };
 
 const DataTableManager = (props: TDataTableManagerProps) => {
+  const { themedValue } = useTheme();
   const areDisplaySettingsEnabled = Boolean(
     props.displaySettings && !props.displaySettings.disableDisplaySettings
   );
@@ -171,7 +174,7 @@ const DataTableManager = (props: TDataTableManagerProps) => {
         onSettingsChange={props.onSettingsChange}
         columnManager={props.columnManager}
         displaySettings={props.displaySettings}
-        managerTheme={props.managerTheme}
+        managerTheme={themedValue(props.managerTheme, 'light')}
       />
       {cloneElement(props.children, {
         columns,