+++ packages/components/collapsible-motion/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @commercetools-uikit/collapsible-motion
 
+## 15.8.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/hooks@15.8.0
+  - @commercetools-uikit/utils@15.8.0
+
+## 15.7.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/hooks@15.7.0
+  - @commercetools-uikit/utils@15.7.0
+
 ## 15.6.0
 
 ### Patch Changes