+++ packages/components/field-errors/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @commercetools-uikit/field-errors
 
+## 15.8.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/messages@15.8.0
+
+## 15.7.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @commercetools-uikit/messages@15.7.0
+
 ## 15.6.0
 
 ### Patch Changes