+++ packages/components/fields/async-select-field/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @commercetools-uikit/async-select-field
 
+## 15.8.0
+
+### Patch Changes
+
+- Updated dependencies [[`806754794`](https://github.com/commercetools/ui-kit/commit/806754794ba1afd0c43370da125d6b4163123814), [`9701afa72`](https://github.com/commercetools/ui-kit/commit/9701afa729b83e5ee0064d0fc12ca73c59c91813)]:
+  - @commercetools-uikit/design-system@15.8.0
+  - @commercetools-uikit/constraints@15.8.0
+  - @commercetools-uikit/field-label@15.8.0
+  - @commercetools-uikit/async-select-input@15.8.0
+  - @commercetools-uikit/spacings@15.8.0
+  - @commercetools-uikit/field-errors@15.8.0
+  - @commercetools-uikit/utils@15.8.0
+
+## 15.7.0
+
+### Patch Changes
+
+- Updated dependencies [[`42d9ea5a2`](https://github.com/commercetools/ui-kit/commit/42d9ea5a251c1fac374c819f17438fbaa254260b), [`3e428cc31`](https://github.com/commercetools/ui-kit/commit/3e428cc31173ac8404bde93b73ff0484ff0330cf), [`5ba1b2bb4`](https://github.com/commercetools/ui-kit/commit/5ba1b2bb486e94981da3a226a72a23957402c3f5), [`8f467aa28`](https://github.com/commercetools/ui-kit/commit/8f467aa283c5888aa55755a2b3e2c549c0d4222a)]:
+  - @commercetools-uikit/design-system@15.7.0
+  - @commercetools-uikit/field-label@15.7.0
+  - @commercetools-uikit/constraints@15.7.0
+  - @commercetools-uikit/async-select-input@15.7.0
+  - @commercetools-uikit/spacings@15.7.0
+  - @commercetools-uikit/field-errors@15.7.0
+  - @commercetools-uikit/utils@15.7.0
+
 ## 15.6.0
 
 ### Patch Changes